Create Charade.cpp
Through this game we tried to implement all those object oriented concepts which we have learned under the guidance of our teacher Mr. Vinay T.R. The game features both single player and 2-players mode and the main aim of the game is to guess the secret number in the least number of attempts.